+++ src/concrete-core/pyproject.toml
@@ -4,17 +4,18 @@ version = "0.0.1"
 description = "AI Agent Orchestration"
 readme = "README.md"
 requires-python = ">=3.11.9, <4.0.0"
+
 dependencies = [
     "python-dotenv==1.0.1",
     "openai==1.40.5",
     "astroid==3.2.3",
-    "astor==0.8.1",
-    "celery==5.4.0",
-    "sqlalchemy==2.0.34",
-    "sqlmodel==0.0.22",
-    "tenacity==9.0.0",
-    "requests==2.32.3",
-    "psycopg[binary]==3.2.3",
+    # "astor==0.8.1",
+    # "celery==5.4.0",
+    # "sqlalchemy==2.0.34",
+    # "sqlmodel==0.0.22",
+    # "tenacity==9.0.0",
+    # "requests==2.32.3",
+    # "psycopg[binary]==3.2.3",
 ]
 
 [build-system]
